requirements.txt
astroid==2.5.1
boto3==1.17.22
botocore==1.20.22
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
idna==2.10
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.2
pyimgur==0.6.0
pylint==2.7.2
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
